refact(lang-monarch): rename typo in package name
