+++ language-monarch/src/main/java/io/github/rosemoe/sora/langs/monarch/registry/model/ThemeModel.kt
@@ -21,9 +21,9 @@
  *     Please contact Rosemoe by email 2073412493@qq.com if you need
  *     additional information or have any questions
  */
-package io.github.rosemoe.sora.langs.monarch.registery.model
+package io.github.rosemoe.sora.langs.monarch.registry.model
 
-import io.github.rosemoe.sora.langs.monarch.registery.FileProviderRegistry
+import io.github.rosemoe.sora.langs.monarch.registry.FileProviderRegistry
 import io.github.rosemoe.sora.langs.monarch.theme.TokenTheme
 import io.github.rosemoe.sora.langs.monarch.theme.toTokenTheme
